# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Fixture Platform API (DEV)
  version: 0.1.0
  description: Phase 1A — User types, Org membership, RBAC modules, Super-admin console.
paths:
...
  /api/orgs/{slug_or_uuid}/:
    get:
      operationId: orgs_retrieve
      description: |-
        GET / PATCH on /api/orgs/{slug_or_uuid}/.

        On GET with a slug match in `SlugRedirect`, returns a 301 with
        Location pointing at the canonical /api/orgs/{uuid}/.
      parameters:
      - in: path
        name: slug_or_uuid
        schema:
          type: string
        required: true
      tags:
      - orgs
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
          description: ''
    patch:
      operationId: orgs_partial_update
      description: |-
        GET / PATCH on /api/orgs/{slug_or_uuid}/.

        On GET with a slug match in `SlugRedirect`, returns a 301 with
        Location pointing at the canonical /api/orgs/{uuid}/.
      parameters:
      - in: path
        name: slug_or_uuid
        schema:
          type: string
        required: true
      tags:
      - orgs
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedOrganizationUpdate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedOrganizationUpdate'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedOrganizationUpdate'
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
          description: ''
...
  /api/orgs/{uuid}:archive/:
    post:
      operationId: orgs_:archive_create
      description: Owner or super-admin only.
      parameters:
      - in: path
        name: uuid
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - orgs
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Archive'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Archive'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Archive'
        required: true
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
          description: ''
  /api/orgs/{uuid}:change_slug/:
    post:
      operationId: orgs_:change_slug_create
      description: POST /api/orgs/{uuid}:change_slug/
      parameters:
      - in: path
        name: uuid
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - orgs
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeSlug'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ChangeSlug'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ChangeSlug'
        required: true
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
          description: ''
  /api/orgs/{uuid}:suspend/:
    post:
      operationId: orgs_:suspend_create
      parameters:
      - in: path
        name: uuid
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - orgs
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Suspend'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Suspend'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Suspend'
        required: true
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
          description: ''
  /api/orgs/{uuid}:transfer_ownership/:
    post:
      operationId: orgs_:transfer_ownership_create
      description: Current owner only. Atomic swap.
      parameters:
      - in: path
        name: uuid
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - orgs
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferOwnership'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/TransferOwnership'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/TransferOwnership'
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
          description: ''
  /api/orgs/{uuid}:unsuspend/:
    post:
      operationId: orgs_:unsuspend_create
      parameters:
      - in: path
        name: uuid
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - orgs
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
          description: ''
...
components:
  schemas:
...
    Archive:
      type: object
      properties:
        reason:
          type: string
          maxLength: 500
      required:
      - reason
...
    ChangeSlug:
      type: object
      properties:
        new_slug:
          type: string
          maxLength: 63
      required:
      - new_slug
...
    Organization:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        slug:
          type: string
          maxLength: 63
        name:
          type: string
          maxLength: 200
        status:
          allOf:
          - $ref: '#/components/schemas/OrganizationStatusEnum'
          readOnly: true
        time_zone:
          type: string
          maxLength: 64
        created_at:
          type: string
          format: date-time
          readOnly: true
        archived_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        suspended_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        suspended_reason:
          type: string
          readOnly: true
      required:
      - archived_at
      - created_at
      - id
      - name
      - slug
      - status
      - suspended_at
      - suspended_reason
...
    OrganizationStatusEnum:
      enum:
      - pending_review
      - active
      - suspended
      - archived
      - orphaned
      type: string
      description: |-
        * `pending_review` - Pending review
        * `active` - Active
        * `suspended` - Suspended
        * `archived` - Archived
        * `orphaned` - Orphaned
...
    PatchedOrganizationUpdate:
      type: object
      properties:
        name:
          type: string
          maxLength: 200
        time_zone:
          type: string
          maxLength: 64
...
    Suspend:
      type: object
      properties:
        reason:
          type: string
          maxLength: 500
      required:
      - reason
    TransferOwnership:
      type: object
      description: |-
        Accepts either ``new_owner_user_id`` (canonical) or ``to_user_id``
        (frontend alias). At least one must be present.
      properties:
        new_owner_user_id:
          type: string
          format: uuid
        to_user_id:
          type: string
          format: uuid
        reason:
          type: string
          maxLength: 500
        event_id:
          type: string
          format: uuid
        conflict_acknowledged:
          type: boolean
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid